fix(character): decimal gold stays readable; get returns worldId; clearer batch array refusal
add_currency stores cents (FINDINGS #111) but the currency schemas still
demanded integers, so one decimal payment made the row fail every read and
broke character list for the whole world. Gold is now decimal in the
character, inventory and corpse schemas.

get returns the row's worldId (null when untagged) so fields: ['worldId']
can check a tag. The batch array refusal names the tool and action, says an
atomic batch rolled back, and points to the direct call.

+++ src/schema/base-schemas.ts
@@ -299,7 +299,8 @@ export type ConditionType = z.infer<typeof ConditionTypeEnum>;
  * Standard D&D currency object
  */
 export const CurrencyFields = z.object({
-    gold: z.number().int().min(0).default(0),
+    gold: z.number().min(0).default(0), // decimal, to the cent
+
     silver: z.number().int().min(0).default(0),
     copper: z.number().int().min(0).default(0),
 }).default({});
